backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
openai
python-multipart
pydantic
openpyxl
python-dotenv

# LINE通知用
line-bot-sdk